Fix dungeon gen for new short arrays

+++ src/minecraft/micdoodle8/mods/galacticraft/moon/wgen/dungeon/GCRoomEmpty.java
@@ -23,7 +23,7 @@ public class GCRoomEmpty extends GCDungeonRoom {
 	}
 
 	@Override
-	public void generate(int[] chunk, int[] meta, int cx, int cz) {
+	public void generate(short[] chunk, byte[] meta, int cx, int cz) {
 		for(int i = posX - 1; i <= posX + sizeX; i++)
 		{
 			for(int j = posY - 1; j <= posY + sizeY; j++)
